+++ src/js/admin/acf-input/index.js
@@ -1,6 +1,7 @@
 import 'acf/field-sweet-spot';
 import 'acf/select-conditions';
 import 'acf/condition-type-regex';
+import 'acf/condition-type-object-type';
 import 'acf/id-field';
 import 'acf/radio-popup';
 import 'acf/select-nav-menu';